Create decoder
This code assumes instruction bits are indexed starting from the right most bit and that the execute and update bits are the right most bits.
I ran the simulation and got the results I expected. The decoder is open to suggestions and changes or if any one wants me to explain in detail just let me know.